Added integer division in scripting language.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@762 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Adventure/AppBuilder/AppBuilder/ScriptFunc.cpp
@@ -744,7 +744,7 @@ int ScriptFunctions::giveLink(ExecutionContext& ctx, unsigned numArgs){
 
 int ScriptFunctions::setNum(ExecutionContext& ctx, unsigned numArgs){
   std::string varname = ctx.stack().pop().getString();
-  int val = ctx.stack().pop().getInt();
+  float val = ctx.stack().pop().getFloat();
   Engine::instance()->getInterpreter()->setVariable(varname, StackData(val));
   return 0;
 }
